Back on track after the horrible pycharm Qt auto-refactoring import mess.
